WIP -> Separated lesson 8 into 2 puzzle groups as a test, working on url navigation logic
